Add improvment to transmit method domain
The field transmit_method_id is displayed twice in the
account.move view depending of the move type.
But distinct domain to each field instances can not be applied
directly on the view (not supported by Odoo)

Using the onchange capability works to a certain extent but when
opening an exsiting record the domain will be incorrect because the
onchange is not executed in that case.

To fix this issue apply the domain at the field declaration with the
help of some new computed fields.